+++ src/cmd/uniq.c
@@ -34,7 +34,7 @@ main(int argc, char *argv[])
 	f = 0;
 	while(argc > 1) {
 		if(*argv[1] == '-') {
-			if(isdigit(argv[1][1]))
+			if(isdigit((uchar)argv[1][1]))
 				fields = atoi(&argv[1][1]);
 			else
 				mode = argv[1][1];
